Refactored indexing to use SPI IncrementalIndexer contracts.
