+++ eZ/Publish/Core/settings/search_engines/common.yml
@@ -26,7 +26,6 @@ parameters:
         ez_document: 'doc'
         ez_fulltext: 'fulltext'
     ezpublish.search.common.field_value_mapper.aggregate.class: eZ\Publish\Core\Search\Common\FieldValueMapper\Aggregate
-    ezpublish.search.common.indexer.class: eZ\Publish\Core\Search\Common\Indexer
 
 services:
     # Note: services tagged with 'ezplatform.field_type.indexable'
